requirements.txt
Django==3.2.7
djangorestframework==3.12.4
drf-spectacular==0.19.0
djangorestframework-simplejwt==4.8.0
django-model-utils==4.1.1
django-filter==21.1
factory-boy==3.2.0
PyPrind==2.11.3
psycopg2==2.9.1
celery==5.1.2
redis==3.5.3